Better plotting

+++ src/03_RPV_modelling.py
@@ -10,6 +10,7 @@ import polars as pl
 from colorama import Fore, Style, init
 from tqdm import tqdm
 
+from src.Util.plotting import plot_df
 from src.Common.config_object import config_object
 from src.Common.rpv import *
 from src.Util.logging import logging_config
@@ -20,6 +21,8 @@ IGNORE_DIRS = {"System Volume Information"}
 PATTERN_TMPL = "*{obj}*.parquet"
 
 init(autoreset=True)
+
+
 
 
 def main():
@@ -40,9 +43,9 @@ def main():
     config = config_object(args.config)
 
     if args.band == 0:
-        band = [f"band{i}" for i in range(1, config.bands + 1)]
+        bands = [f"band{i}" for i in range(1, config.bands + 1)]
     else:
-        band = [args.band]
+        bands = [args.band]
 
     # Search data
     folders = ["", "metashape", "products_uav_data", "output", "extract", "polygon_df"]
@@ -67,10 +70,18 @@ def main():
 
     # Create rpvs for each
     for week, gdf in weeks_dics.items():
-        for band in band:
-            result = process_weekly_data({week: gdf}, band=band)
-            out_dir = Path(base_dir) / "RPV_Results" / "V6"
+        for band in bands:
+            out_dir = Path(base_dir) / "RPV_Results" / "V8"
             out_dir.mkdir(parents=True, exist_ok=True)
+
+
+            plot_df(week,gdf,band)
+
+            if (out_dir / f"rpv_{week}_{band}_results.csv").exists():
+                continue
+
+
+            result = process_weekly_data({week: gdf}, band=band,sample_total_dataset=500_000)
             result.drop("geometry").write_csv(str(out_dir / f"rpv_{week}_{band}_results.csv"))
